requirements.txt
Flask
flask-bcrypt
psycopg2-binary
pytest
selenium
webdriver-manager
